requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
chromadb>=0.5.0
openai>=1.58.1,<2.0.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pypdf>=4.0.2
